Refactored hal vulkan into multiple files, added whitespace across hal, moved every backend to its own folder
