+++ crates/supasim-hal/src/metal/mod.rs
@@ -30,27 +30,34 @@ use crate::{
 };
 
 struct UniqueObject<T: ?Sized + NSObjectProtocol>(Retained<ProtocolObject<T>>);
+
 unsafe impl<T: ?Sized + NSObjectProtocol> Send for UniqueObject<T> {}
+
 impl<T: ?Sized + NSObjectProtocol> UniqueObject<T> {
     pub fn new(inner: Retained<ProtocolObject<T>>) -> Self {
         Self(inner)
     }
+
     #[allow(dead_code)]
     pub fn into_inner(self) -> Retained<ProtocolObject<T>> {
         self.0
     }
+
     /// # Safety
     /// This should only be called if the cloned value will automatically be synced or won't be used.
     pub unsafe fn unsafe_clone(&self) -> Self {
         Self(self.0.clone())
     }
 }
+
 impl<T: ?Sized + NSObjectProtocol> Deref for UniqueObject<T> {
     type Target = ProtocolObject<T>;
+
     fn deref(&self) -> &ProtocolObject<T> {
         &self.0
     }
 }
+
 impl<T: ?Sized + NSObjectProtocol> std::fmt::Debug for UniqueObject<T> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         self.0.fmt(f)
@@ -59,6 +66,7 @@ impl<T: ?Sized + NSObjectProtocol> std::fmt::Debug for UniqueObject<T> {
 
 #[derive(Debug, Clone)]
 pub struct Metal;
+
 impl Metal {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     pub fn create_instance() -> Result<MetalInstance, MetalError> {
@@ -73,11 +81,13 @@ impl Metal {
 pub struct MetalDevice {
     device: UniqueObject<dyn MTLDevice>,
 }
+
 impl Device<Metal> for MetalDevice {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn cleanup_cached_resources(&self, _instance: &MetalInstance) -> Result<(), MetalError> {
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn create_buffer(
         &self,
@@ -101,6 +111,7 @@ impl Device<Metal> for MetalDevice {
             buffer: UniqueObject::new(buffer),
         })
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn import_buffer(
         &self,
@@ -108,6 +119,7 @@ impl Device<Metal> for MetalDevice {
     ) -> Result<<Metal as Backend>::Buffer, <Metal as Backend>::Error> {
         unreachable!()
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn create_semaphore(
         &self,
@@ -121,6 +133,7 @@ impl Device<Metal> for MetalDevice {
             current_value: 0,
         })
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn import_semaphore(
         &self,
@@ -128,6 +141,7 @@ impl Device<Metal> for MetalDevice {
     ) -> Result<<Metal as Backend>::Semaphore, <Metal as Backend>::Error> {
         unreachable!()
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     fn get_properties(&self, _instance: &MetalInstance) -> types::HalDeviceProperties {
         let is_unified_memory = self.device.hasUnifiedMemory();
@@ -139,6 +153,7 @@ impl Device<Metal> for MetalDevice {
             supports_semaphore_import: false,
         }
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn destroy(
         self,
@@ -152,6 +167,7 @@ impl Device<Metal> for MetalDevice {
 pub struct MetalStream {
     inner: UniqueObject<dyn MTLCommandQueue>,
 }
+
 impl Stream<Metal> for MetalStream {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn create_bind_group(
@@ -172,6 +188,7 @@ impl Stream<Metal> for MetalStream {
             buffers: Mutex::new(buffers),
         })
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn create_recorder(
         &self,
@@ -182,6 +199,7 @@ impl Stream<Metal> for MetalStream {
             command_buffer: UniqueObject::new(command_buffer),
         })
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn submit_recorders(
         &mut self,
@@ -204,6 +222,7 @@ impl Stream<Metal> for MetalStream {
         }
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn wait_for_idle(&mut self) -> Result<(), <Metal as Backend>::Error> {
         let cb = self.inner.commandBuffer().ok_or(MetalError::ObjectCreate)?;
@@ -211,6 +230,7 @@ impl Stream<Metal> for MetalStream {
         cb.waitUntilCompleted();
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn cleanup_cached_resources(
         &self,
@@ -218,6 +238,7 @@ impl Stream<Metal> for MetalStream {
     ) -> Result<(), <Metal as Backend>::Error> {
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn destroy(
         self,
@@ -274,6 +295,7 @@ pub struct MetalInstance {
     // the inner type to be send + sync
     device: UniqueObject<dyn MTLDevice>,
 }
+
 impl BackendInstance<Metal> for MetalInstance {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     fn get_properties(&self) -> HalInstanceProperties {
@@ -298,6 +320,7 @@ impl BackendInstance<Metal> for MetalInstance {
             atomic_int64: false,
         }
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn create_semaphore(
         &self,
@@ -311,6 +334,7 @@ impl BackendInstance<Metal> for MetalInstance {
             current_value: 0,
         })
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn compile_kernel(
         &self,
@@ -354,10 +378,12 @@ impl BackendInstance<Metal> for MetalInstance {
             revised_buffer_indices: revised_layout,
         })
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn cleanup_cached_resources(&mut self) -> Result<(), <Metal as Backend>::Error> {
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn destroy(self) -> Result<(), <Metal as Backend>::Error> {
         Ok(())
@@ -382,6 +408,7 @@ pub enum MetalError {
     #[error("An invalid utf8 string was passed to compile_kernel")]
     NonUtf8KernelString,
 }
+
 impl crate::Error<Metal> for MetalError {
     fn is_out_of_device_memory(&self) -> bool {
         false
@@ -399,6 +426,7 @@ pub struct MetalSemaphore {
     event: UniqueObject<dyn MTLSharedEvent>,
     current_value: u64,
 }
+
 impl Semaphore<Metal> for MetalSemaphore {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn is_signalled(
@@ -408,16 +436,19 @@ impl Semaphore<Metal> for MetalSemaphore {
         let value = self.event.signaledValue();
         Ok(value == self.current_value + 1)
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn reset(&mut self, _device: &MetalInstance) -> Result<(), <Metal as Backend>::Error> {
         self.current_value += 1;
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn signal(&mut self, _device: &MetalInstance) -> Result<(), <Metal as Backend>::Error> {
         self.event.setSignaledValue(self.current_value + 1);
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn wait(&self, _device: &MetalInstance) -> Result<(), <Metal as Backend>::Error> {
         self.event
@@ -435,6 +466,7 @@ pub struct MetalBindGroup {
     #[allow(clippy::type_complexity)]
     buffers: Mutex<Vec<(UniqueObject<dyn MTLBuffer>, u64, u64)>>,
 }
+
 impl BindGroup<Metal> for MetalBindGroup {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn update(
@@ -454,6 +486,7 @@ impl BindGroup<Metal> for MetalBindGroup {
         }
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn destroy(
         self,
@@ -474,6 +507,7 @@ pub struct MetalKernel {
     /// in the kernel's reordered layout
     revised_buffer_indices: Vec<usize>,
 }
+
 impl Kernel<Metal> for MetalKernel {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn destroy(
@@ -488,6 +522,7 @@ impl Kernel<Metal> for MetalKernel {
 pub struct MetalBuffer {
     buffer: UniqueObject<dyn MTLBuffer>,
 }
+
 impl Buffer<Metal> for MetalBuffer {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn map(
@@ -497,6 +532,7 @@ impl Buffer<Metal> for MetalBuffer {
         assert!((self.buffer.storageMode().0 & MTLResourceOptions::StorageModePrivate.0) == 0);
         Ok(self.buffer.contents().as_ptr() as *mut u8)
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn read(
         &mut self,
@@ -510,6 +546,7 @@ impl Buffer<Metal> for MetalBuffer {
         data.copy_from_slice(slice);
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn write(
         &mut self,
@@ -523,6 +560,7 @@ impl Buffer<Metal> for MetalBuffer {
         slice.copy_from_slice(data);
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn unmap(
         &mut self,
@@ -530,6 +568,7 @@ impl Buffer<Metal> for MetalBuffer {
     ) -> Result<(), <Metal as Backend>::Error> {
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn destroy(
         self,
@@ -545,6 +584,7 @@ enum CurrentEncoder {
     Blit(Retained<ProtocolObject<dyn MTLBlitCommandEncoder>>),
     Compute(Retained<ProtocolObject<dyn MTLComputeCommandEncoder>>),
 }
+
 impl CurrentEncoder {
     fn blit(
         &mut self,
@@ -563,6 +603,7 @@ impl CurrentEncoder {
             }
         })
     }
+
     fn compute(
         &mut self,
         cb: &ProtocolObject<dyn MTLCommandBuffer>,
@@ -582,6 +623,7 @@ impl CurrentEncoder {
             }
         })
     }
+
     fn finish(&mut self) {
         match self {
             Self::Blit(b) => b.endEncoding(),
@@ -596,12 +638,14 @@ impl CurrentEncoder {
 pub struct MetalCommandRecorder {
     command_buffer: UniqueObject<dyn MTLCommandBuffer>,
 }
+
 impl CommandRecorder<Metal> for MetalCommandRecorder {
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn clear(&mut self, stream: &MetalStream) -> Result<(), <Metal as Backend>::Error> {
         *self = unsafe { stream.create_recorder()? };
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn record_commands(
         &mut self,
@@ -693,6 +737,7 @@ impl CommandRecorder<Metal> for MetalCommandRecorder {
         encoder.finish();
         Ok(())
     }
+
     #[cfg_attr(feature = "trace", tracing::instrument)]
     unsafe fn destroy(
         self,
